Test automation/Playwright/Playwright_Pytest_OrangeHRM/requirements.txt
# Minimal dependencies - automatically installs all sub-dependencies
# Test Framework
pytest>=7.0.0
pytest-html>=3.1.0
pytest-rerunfailures>=12.0
playwright>=1.40.0

# Utilities
PyYAML>=6.0
